Fixed the type of some ivar access functions.  Ivars have type uintptr_t, not u_long.  On i386's with 64-bit longs, returning u_longs indirectly in (more than) the space reserved for uintptr_t's tended to corrupt the previous frame pointer in the stack frame, so it was not easy to debug. The type mismatches are hidden by the bogus cast in DEVMETHOD().

+++ sys/pci/pcisupport.c
@@ -769,7 +769,7 @@ static int pcib_attach(device_t dev)
 }
 
 static int
-pcib_read_ivar(device_t dev, device_t child, int which, u_long *result)
+pcib_read_ivar(device_t dev, device_t child, int which, uintptr_t *result)
 {
 	if (which == PCIB_IVAR_HOSE) {
 		/*
